+++ tests/test_visitors/test_tokenize/test_primitives/conftest.py
@@ -19,5 +19,5 @@ statement_with_expression = 'other_var + {0}'
     statement_with_expression,
 ])
 def primitives_usages(request):
-    """Fixture to return possible cases of promitives use cases."""
+    """Fixture to return possible cases of primitives use cases."""
     return request.param